No Unwnowns (#39)
* Unknown Type

Much of this work is turning unknown into an error. And all the resultant
bugs that needed fixing. Lot's of good type stuff in here.

* Just a lot of bug fixes related to no unknowns

* Fix Cargo.toml to not point at local sarzak.

* Fixed proxy objects to have impls.

* Test fixing.

* Test fixing.

* Locals in match expressions ++

The big deal is that now match expressions can re-use the
variable name being matched against inside of the match arms.

Like so:

```rust
let result = do_something();
match result {
    Ok(result) => {
        // result is the result of the Ok variant
    }
    Err(result) => {
        // result is the result of the Err variant
    }
}
````

* Really fixed that match issue

The fix I did broke other stuff. Specifically, a test where
a variable was reused in the test and it had different types.
The test failed because I was only grabbing the first type.

The real fix is to reverse the order of the values being pulled
from the blocks. But all of the values in the block need to be
maintained.

There is also some map related stuff in here. I think getting
map working was where this all started...

* Remove trace during VM tests.

* Added list.join

There is also some output and cleanup of the code.

* Got Map::get working

Hashmaps are still limited to string keys. Not sure about
values.

* HashMap parsing

Now parsing HashMap<_, _> as a type.

* Added dockerfiles

It's cute, I guess.

* Added test

* tests

* Added Map get test

Uncovered a few bugs too. Go tests!

* Added Map get test

Uncovered a few bugs too. Go tests!

* map insert test

* for loop iterator testing

* Moar tests

* Testing

* Added list.map to the VM

More tests as well. In fact, it was a map test that
brought it to my attention that map wasn't in the VM.

* More coverage for list, and anylist.

* trying to eeke out more coverage.

* Markdown theme support

Now I can change the mermaid theme in the markdown plugin.

* Fix md test. I really need to check func signatures sooner than later...

* Coverage! 🚀