Whoops -- too much.
There is a fix for the sqlx plugin that specifies the type of
the thing being bound.

There is a fix for blessing tests. I needed to disable all the
noise that the extruder and compiler make.

There is an async block addition too. Now you can create async
blocks and await them:

```rust
fn main() {
    let x = async {
        let y = 1;
        y
    };
    let z = x.await;
    // This will print 1
    println!("{}", z);
}
```
`