Working on FFI tests
Ran into this thing with FFi lists not having a type.
Sorted it out by giving it the option of a type.
Fixed the into value trait impl that's used after the
invoke_func call to use the type from the ffi list and
not the thing we pass in, which is a result.