Website compiles -- eventually
Fixed a bug where we'd find a type in the store but it wasn't being
added to the context's types and thus was being marked as not found.

Updated outdated crates. Snafu changed it's interface a bit and it
no longer requires the `Location` import. That touched a lot of files.

Fixed an FFI bug with the plugins where a `Result` was being typed
as `Result` and not `std::result::Result`.